trinethra/services/automation_api/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
pydantic==2.10.6
python-dotenv==1.0.1
requests==2.32.3
psycopg2-binary==2.9.9